add an event handler for the elasticsearch-relation changed event
The graylog docs say that server.conf requires a field called 'elasticsearch_hosts' which is list of _some_ of the ES hosts on the cluster. In this iteration, we are just using one address - the ingress address of the whole ES cluster. This will need testing, but I suspect this will work.

The major remaining TODO before the MVP charm can be up and running is to add the mongodb relation.