fix the bind_address property to use hostname rather than 'network-get' output
We were running into an issue where Graylog was failing to start because of a misconfigured bind address. We need to use a host name because an IP will be invalidated as soon as it is set in the Juju pod spec, which will trigger a pod restart and then the IP of the unit will change.