add http_external_uri, associated charm property, and readiness probe
There was an issue where the web interface was not loading. It turned out that the default http_external_uri was the same as the bind address. To work, this field needs to refer to the application's ingress-address.

This commit also adds a readiness probe to Kubernetes. It seems to take a while to actually return properly, but the web interface is fully functional as of this commit.