First working version of Graylog operator (#16)
* boilerplate commit

* add basic config file with port option

* add basic metadata.yaml

* add basic pod-spec configuration

* add starting README

* add OCI image resource to metadata.yaml

* add specific oci-image commit to requirements.txt

* use resources rather than hardcoded configuration for the graylog image in the charm

* Create python-package.yaml

* Remove trailing whitespace

* Add develop to workflow

* Install virtualenv

* Change python versions

* Trying to run test with github actions

* Debuging GA

* Typo!

* Create python-package.yaml

* Remove trailing whitespace

* Add develop to workflow

* Install virtualenv

* Change python versions

* Trying to run test with github actions

* Debuging GA

* Typo!

* Add admin-password as a config option because it is _required_ by Graylog to be configured via the server.conf file.

There may be a better way to do this in the future, but a config option might be a decent choice for now.

* add the elasticsearch relation to metadata.yaml

* add an event handler for the elasticsearch-relation changed event

The graylog docs say that server.conf requires a field called 'elasticsearch_hosts' which is list of _some_ of the ES hosts on the cluster. In this iteration, we are just using one address - the ingress address of the whole ES cluster. This will need testing, but I suspect this will work.

The major remaining TODO before the MVP charm can be up and running is to add the mongodb relation.

* add two tests to reflect the elasticsearch relation

Three changes:
1) change 'test_pod_spec_for_leader' to 'test_pod_spec_port' and use the harness's 'get_pod_spec' function rather that directly calling 'charm._build_pod_spec'
2) add a test to determine whether the _stored variable contains the correct elasticsearch URI
3) add a test to check the actual contents of the server.conf config file.

* add __pycache__ to gitignore

* add limit of 1 to elasticsearch relation in metadata.yaml

* add elasticsearch_relation_broken handler

When a relation is broken, we want to remove the elasticsearch_uri data from _stored and reconfigure the pod. If we don't Graylog might end up trying to read from a non-existent or unreachable ES cluster.

* FIX #7 - Run our test in python 3.5 too

* add a few assertions to the existing tests

1) check the has_elasticsearch property
2) check that _stored.elasticsearch_uri is cleared when we have a relation broken event.

* Add admin-password as a config option because it is _required_ by Graylog to be configured via the server.conf file.

There may be a better way to do this in the future, but a config option might be a decent choice for now.

* add the elasticsearch relation to metadata.yaml

* add an event handler for the elasticsearch-relation changed event

The graylog docs say that server.conf requires a field called 'elasticsearch_hosts' which is list of _some_ of the ES hosts on the cluster. In this iteration, we are just using one address - the ingress address of the whole ES cluster. This will need testing, but I suspect this will work.

The major remaining TODO before the MVP charm can be up and running is to add the mongodb relation.

* add two tests to reflect the elasticsearch relation

Three changes:
1) change 'test_pod_spec_for_leader' to 'test_pod_spec_port' and use the harness's 'get_pod_spec' function rather that directly calling 'charm._build_pod_spec'
2) add a test to determine whether the _stored variable contains the correct elasticsearch URI
3) add a test to check the actual contents of the server.conf config file.

* add __pycache__ to gitignore

* add limit of 1 to elasticsearch relation in metadata.yaml

* add elasticsearch_relation_broken handler

When a relation is broken, we want to remove the elasticsearch_uri data from _stored and reconfigure the pod. If we don't Graylog might end up trying to read from a non-existent or unreachable ES cluster.

* add a few assertions to the existing tests

1) check the has_elasticsearch property
2) check that _stored.elasticsearch_uri is cleared when we have a relation broken event.

* Temporary: point oci-image requirement to a personal repo while upstream waits for PR to land

* REL #10 - I add event observations to mongodb

* REL #10 - [WIP] Add handlers..

* add mongodb relation to metadata.yaml

* fix typo in config.yaml

option -> options

* specify statefulset deployment in metadata.yaml

When deploying as a stateless app, the pods were unnecessarily restarting if there was ever a configuration change.

* Remove test that checked server.conf contents

Now that we are using environmental variables for configuration, we have no need for the test.

* Add mongodb relation handling and configuration via environmental variables

*NOTE* this charm is not in a working state at this point and is running into an error (when deployed with ES and Mongo) that looks like this:

org.graylog2.shared.initializers.InputSetupService - Not starting any inputs because lifecycle is: Failed [LB:DEAD]

* fix the bind_address property to use hostname rather than 'network-get' output

We were running into an issue where Graylog was failing to start because of a misconfigured bind address. We need to use a host name because an IP will be invalidated as soon as it is set in the Juju pod spec, which will trigger a pod restart and then the IP of the unit will change.

* add http_external_uri, associated charm property, and readiness probe

There was an issue where the web interface was not loading. It turned out that the default http_external_uri was the same as the bind address. To work, this field needs to refer to the application's ingress-address.

This commit also adds a readiness probe to Kubernetes. It seems to take a while to actually return properly, but the web interface is fully functional as of this commit.

* add a patch for 'external_uri' property

Since this property only works in the context of Juju, we need to patch it. We will need to add specific return values to the mocked property if we want to directly test anything that includes a reference to this property.

Co-authored-by: Jose C. Massón <jose.masson@canonical.com>
Co-authored-by: Jose C. Massón <jose.masson@canonical.coop>